[migrate] move decorators to original modules for Git diff [migrate] move WebField, AsyncBox & utilities to Legacy folder for now [remove] Renderer module
