+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "web-cell",
-    "version": "3.0.0-rc.1",
+    "version": "3.0.0-rc.2",
     "description": "Web Components engine based on VDOM, JSX, MobX & TypeScript",
     "keywords": [
         "web",
@@ -27,7 +27,7 @@
     "types": "dist/index.d.ts",
     "dependencies": {
         "@swc/helpers": "^0.5.3",
-        "dom-renderer": "^2.0.1",
+        "dom-renderer": "^2.0.2",
         "mobx": ">=6.11",
         "regenerator-runtime": "^0.14.1",
         "web-utility": "^4.1.3"
@@ -54,7 +54,7 @@
         "husky": "^8.0.3",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
-        "jsdom": "^23.0.1",
+        "jsdom": "^23.1.0",
         "lint-staged": "^15.2.0",
         "open-cli": "^8.0.0",
         "parcel": "~2.11.0",
@@ -73,7 +73,7 @@
         "pack-preview": "cd preview/  &&  rimraf ../.parcel-cache dist/  &&  parcel build --public-url=. --dist-dir=../docs/preview/",
         "pack-dist": "parcel build source/index.ts",
         "pack-docs": "rimraf docs/  &&  typedoc source/",
-        "build": "rimraf dist/  &&  npm run pack-dist  &&  npm run pack-preview",
+        "build": "rimraf dist/  &&  npm run pack-dist  &&  npm run pack-docs  &&  npm run pack-preview",
         "start": "npm run pack-docs  &&  open-cli docs/index.html",
         "prepublishOnly": "npm test  &&  npm run build"
     },
