+++ source/WebCell.tsx
@@ -1,172 +1,134 @@
-import type {} from 'element-internals-polyfill';
-import { ElementInternals } from 'element-internals-polyfill/dist/element-internals';
+import { DOMRenderer, VNode } from 'dom-renderer';
 import {
-    Constructor,
     CustomElement,
-    CustomElementClass,
+    DelegateEventHandler,
     delegate,
-    stringifyDOM,
-    toCamelCase
+    stringifyDOM
 } from 'web-utility';
-import { IReactionDisposer, observable } from 'mobx';
 
-import { WebCellProps, getMobxData } from './utility';
-import {
-    ComponentMeta,
-    DOMEventDelegater,
-    ReactionDelegater
-} from './decorator';
-import { Fragment, createCell, render } from './renderer';
-
-export interface WebCellComponent<P extends WebCellProps = WebCellProps>
-    extends CustomElement {
-    internals?: ElementInternals;
-    root?: DocumentFragment | HTMLElement;
-    update?(): void;
-    /**
-     * @deprecated Get values from `this` directly since WebCell 3.0.0
-     */
-    props?: P;
-    disposers?: IReactionDisposer[];
-    syncPropAttr?(name: string): void;
-    defaultSlot?: JSX.Element;
-    render?(): JSX.Element | undefined;
-    /**
-     * Called after rendering
-     */
-    updatedCallback?(): void;
-    emit?(event: string, detail?: any, options?: EventInit): boolean;
-    toString(): string;
+export interface ComponentMeta
+    extends ElementDefinitionOptions,
+        Partial<ShadowRootInit> {
+    tagName: string;
 }
 
-export interface WebCellClass<P extends WebCellProps = WebCellProps>
-    extends Pick<CustomElementClass, 'observedAttributes'>,
-        Partial<ComponentMeta>,
-        Constructor<WebCellComponent<P>> {
-    eventDelegaters?: DOMEventDelegater[];
-    reactions?: ReactionDelegater[];
+export type ComponentClass = CustomElementConstructor;
+
+interface DelegatedEvent {
+    type: keyof HTMLElementEventMap;
+    selector: string;
+    handler: EventListener;
 }
+const eventMap = new WeakMap<CustomElement, DelegatedEvent[]>();
 
-export function WebCell<P extends WebCellProps = WebCellProps>(
-    superClass: Constructor<CustomElement> = HTMLElement
-): WebCellClass<P> {
-    class WebCell extends superClass implements WebCellComponent<P> {
-        static tagName: ComponentMeta['tagName'];
-        static extends?: ComponentMeta['extends'];
-        static mode?: ComponentMeta['mode'];
-        static delegatesFocus?: ComponentMeta['delegatesFocus'];
-        static eventDelegaters: DOMEventDelegater[] = [];
-        static reactions: ReactionDelegater[] = [];
+/**
+ * `class` decorator of Web components
+ */
+export function component(meta: ComponentMeta) {
+    return <T extends ComponentClass>(
+        Class: T,
+        { addInitializer }: ClassDecoratorContext<ComponentClass>
+    ) => {
+        class RendererComponent
+            extends (Class as ComponentClass)
+            implements CustomElement
+        {
+            protected internals = this.attachInternals();
+            protected renderer = new DOMRenderer();
 
-        readonly internals?: ElementInternals;
-        readonly root: DocumentFragment | HTMLElement;
+            get root() {
+                return this.internals.shadowRoot || this;
+            }
 
-        get props() {
-            return getMobxData<P>(this);
-        }
-        readonly disposers: IReactionDisposer[] = [];
+            constructor() {
+                super();
 
-        @observable
-        defaultSlot?: JSX.Element;
+                if (meta.mode && !this.internals.shadowRoot)
+                    this.attachShadow(meta as ShadowRootInit);
+            }
 
-        [key: string]: any;
+            connectedCallback() {
+                this.update();
 
-        constructor() {
-            super();
+                const { mode } = meta;
+                const renderChildren = !(mode != null);
 
-            const {
-                extends: extendTag,
-                mode,
-                delegatesFocus,
-                eventDelegaters
-            } = this.constructor as WebCellClass;
+                const { root } = this,
+                    events = eventMap.get(this) || [];
 
-            const renderChildren = !(mode != null);
+                for (const { type, selector, handler } of events) {
+                    if (renderChildren && /^:host/.test(selector))
+                        console.warn(
+                            `[WebCell] DOM Event delegation of "${selector}" won't work if you don't invoke "this.attachShadow()" manually.`
+                        );
+                    root.addEventListener(type, handler);
+                }
 
-            if (!extendTag) this.internals = this.attachInternals();
+                super['connectedCallback']?.();
+            }
 
-            this.root = renderChildren
-                ? this
-                : this.internals?.shadowRoot ||
-                  this.attachShadow({ mode, delegatesFocus });
+            declare render: () => VNode;
 
-            for (const { selector, method } of eventDelegaters) {
-                if (renderChildren && /^:host/.test(selector))
-                    console.warn(
-                        `[WebCell] DOM Event delegation of "${selector}" won't work if you don't invoke "this.attachShadow()" manually.`
-                    );
-                this[method] = delegate(selector, this[method]).bind(this);
+            update() {
+                const vNode = this.render?.();
+
+                if (vNode) this.renderer.render(vNode, this.root);
+            }
+
+            disconnectedCallback() {
+                const { root } = this,
+                    events = eventMap.get(this) || [];
+
+                for (const { type, handler } of events)
+                    root.removeEventListener(type, handler);
+
+                super['disconnectedCallback']?.();
+            }
+
+            emit(
+                event: string,
+                detail?: any,
+                { cancelable, bubbles, composed }: EventInit = {}
+            ) {
+                return this.dispatchEvent(
+                    new CustomEvent(event, {
+                        detail,
+                        cancelable,
+                        bubbles,
+                        composed
+                    })
+                );
+            }
+
+            toString() {
+                return stringifyDOM(this.root);
             }
         }
 
-        connectedCallback() {
-            const { eventDelegaters } = this.constructor as WebCellClass,
-                { root } = this;
-
-            for (const { type, method } of eventDelegaters)
-                root.addEventListener(type, this[method]);
-
-            if (!(root instanceof DocumentFragment) || !root.lastChild)
-                this.update();
-        }
-
-        render() {
-            return (this.constructor as WebCellClass).mode != null ? (
-                <slot />
-            ) : (
-                this.defaultSlot
-            );
-        }
-
-        update() {
-            render(this.render() ?? <></>, this.root);
-
-            this.updatedCallback?.();
-        }
-
-        disconnectedCallback() {
-            const { eventDelegaters } = this.constructor as WebCellClass;
-
-            for (const { type, method } of eventDelegaters)
-                this.root.removeEventListener(type, this[method]);
-
-            for (const disposer of this.disposers) disposer();
-
-            this.disposers.length = 0;
-        }
-
-        syncPropAttr(name: string) {
-            const value = this[toCamelCase(name)];
-
-            if (value != null && value !== false) {
-                if (typeof value !== 'object')
-                    super.setAttribute(name, value === true ? name : value);
-            } else this.removeAttribute(name);
-        }
-
-        emit(
-            event: string,
-            detail?: any,
-            { cancelable, bubbles, composed }: EventInit = {}
-        ) {
-            return this.dispatchEvent(
-                new CustomEvent(event, {
-                    detail,
-                    cancelable,
-                    bubbles,
-                    composed
-                })
-            );
-        }
-
-        toString() {
-            return stringifyDOM(this.root);
-        }
-    }
-
-    return WebCell;
+        addInitializer(function () {
+            globalThis.customElements?.define(meta.tagName, this, meta);
+        });
+        return RendererComponent as unknown as T;
+    };
 }
 
-export type ComponentClass = {
-    new (): InstanceType<ReturnType<typeof WebCell>>;
-};
+/**
+ * Method decorator of DOM Event delegation
+ */
+export function on<T extends HTMLElement>(
+    type: DelegatedEvent['type'],
+    selector: string
+) {
+    return (
+        method: DelegateEventHandler,
+        { addInitializer }: ClassMethodDecoratorContext<T>
+    ) =>
+        addInitializer(function () {
+            const events = eventMap.get(this) || [],
+                handler = delegate(selector, method.bind(this));
+
+            events.push({ type, selector, handler });
+
+            eventMap.set(this, events);
+        });
+}
